Fixed compile errors on windows with call to sqrt with integer parameters.
git-svn-id: https://gforge.sci.utah.edu/svn/nektar/trunk@3394 305cdda6-5ce1-45b3-a98d-dfc68c8b3305

+++ library/SpatialDomains/QuadGeom.cpp
@@ -167,7 +167,7 @@ namespace Nektar
             for(int i = 0; i < m_coordim; ++i)
             {
                 int npts = curve->m_points.size();
-                int nEdgePts = (int)sqrt(npts);
+                int nEdgePts = (int)sqrt(static_cast<double>(npts));
                 Array<OneD,NekDouble> tmp(npts);
                 LibUtilities::PointsKey curveKey(nEdgePts, curve->m_ptype);
